requirements-dev.txt
requests==2.15.0
pytest==4.3.1
mock==2.0.0
python-socketio==4.2.0